requirements.txt
praw~=7.1.0
pandas~=1.1.1
python-dotenv~=0.14.0
psaw